extract useUpdateTilePropsSaver to a hooks folder
